function_app/ingest/requirements.txt
azure-functions==1.21.3
azure-identity==1.19.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.23.1
azure-storage-queue==12.10.0
azure-data-tables==12.6.0
azure-search-documents==11.6.0
azure-ai-documentintelligence==1.0.0
openai==1.55.3
msal==1.31.1
requests==2.32.3
pypdf==4.3.1
